fix: toolbar buttons
