+++ lib/common/navigation/app_bars/editor_app_bar.dart
@@ -7,7 +7,7 @@ import 'package:localmaterialnotes/common/actions/pin.dart';
 import 'package:localmaterialnotes/common/actions/restore.dart';
 import 'package:localmaterialnotes/common/navigation/menu_options.dart';
 import 'package:localmaterialnotes/models/note/note.dart';
-import 'package:localmaterialnotes/pages/editor/about_sheet.dart';
+import 'package:localmaterialnotes/pages/editor/sheets/about_sheet.dart';
 import 'package:localmaterialnotes/providers/notifiers.dart';
 import 'package:localmaterialnotes/utils/constants/constants.dart';
 import 'package:localmaterialnotes/utils/constants/paddings.dart';
