+++ lib/pages/editor/editor_page.dart
@@ -8,8 +8,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:localmaterialnotes/common/placeholders/loading_placeholder.dart';
 import 'package:localmaterialnotes/common/routing/router.dart';
 import 'package:localmaterialnotes/models/note/note.dart';
-import 'package:localmaterialnotes/pages/editor/editor_field.dart';
-import 'package:localmaterialnotes/pages/editor/editor_toolbar.dart';
+import 'package:localmaterialnotes/pages/editor/widgets/editor_field.dart';
+import 'package:localmaterialnotes/pages/editor/widgets/editor_toolbar.dart';
 import 'package:localmaterialnotes/providers/notes/notes_provider.dart';
 import 'package:localmaterialnotes/providers/notifiers.dart';
 import 'package:localmaterialnotes/utils/constants/constants.dart';
